+++ docs/changes.md
@@ -13,6 +13,7 @@ Release date: `2018-??-??`
 #### Minor changes
 - Doc: Add a link to know up-to-date envvars value
 - Framework: Do not create test folder for case sensitivity in the Drive folder but in a temporary one
+- Framework: Review DirectEdit class
 - Framework: Review EngineNext class
 - Jenkins: Discard old builds
 - Tests: Add a simple test for .rvt files